backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Repository Management
GitPython>=3.1.40

# Caching
redis>=5.0.0

# Property-Based Testing
hypothesis>=6.92.0

# Background Processing
celery>=5.3.0
celery[redis]>=5.3.0
